wrapping in int() to suppress deprecation

+++ acq/scanimage/ephusRoutines.py
@@ -117,8 +117,8 @@ def parseXSG(filename):
                         stim_array = np.zeros(sampleRate*traceLength) + offset
 
                         for pulse_number in range(number_of_pulses):
-                            start = pulse_number * ISI + delay
-                            end = start + width
+                            start = int(pulse_number * ISI + delay)
+                            end = int(start + width)
                             stim_array[start:end] = amp
                         if on :
                             xsgDict['stimulator'][header['stimulator']['stimulator']['channels']['channelName'][pulse]] = stim_array
